jobs/backend/requirements.txt
appdirs==1.4.4
APScheduler==3.7.0
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
idna==2.10
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
mysql-connector-python-rf==2.2.2
pathspec==0.8.1
pycodestyle==2.6.0
pylint==2.6.0
python-dateutil==2.8.1
pytz==2020.5
regex==2020.11.13
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.3
Werkzeug==1.0.1
wrapt==1.12.1
yapf==0.30.0
